backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
requests==2.32.3

# AI/ML dependencies
openai==1.54.3
# PyMuPDF==1.23.7  # Comentado - requer Visual Studio Build Tools no Windows
redis==5.0.4
